pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = 'electronicparsers'
version = '1.1'
description = 'Collection of NOMAD parsers for electronic structure codes.'
readme = "README.md"
authors = [{ name = "The NOMAD Authors" }]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab[infrastructure]@git+https://github.com/nomad-coe/nomad.git@develop",
    "nomad-schema-plugin-simulation-workflow@git+https://github.com/nomad-coe/nomad-schema-plugin-simulation-workflow.git@develop",
    "nomad-schema-plugin-run@git+https://github.com/nomad-coe/nomad-schema-plugin-run.git@develop",
    "simulationparsers@git+https://github.com/nomad-coe/simulation-parsers.git@develop",
    "lxml==4.7.1",
    "netCDF4==1.5.4",
    "h5py>=3.6.0",
    "pyyaml==6.0",
]

[project.urls]
homepage = "https://github.com/nomad-coe/electronic-parsers"

[project.optional-dependencies]
dev = [
    'mypy==1.0.1',
    'pytest==3.10.0',
    'pytest-timeout==1.4.2',
    'pytest-cov==2.7.1',
    'astroid==2.11.7',
    'typing-extensions==4.4.0',
    'ruff==0.1.8'
]

[tool.ruff]
include = ["electronicparsers/*.py", "tests/*.py"]
select = [
    "E", # pycodestyle
    "W", # pycodestyle
    "PL", # pylint
]
ignore = [
	"E501", # Line too long ({width} > {limit} characters)
	"E701", # Multiple statements on one line (colon)
	"E731", # Do not assign a lambda expression, use a def
	"E402", # Module level import not at top of file
	"PLR0911", # Too many return statements
	"PLR0912", # Too many branches
	"PLR0913", # Too many arguments in function definition
	"PLR0915", # Too many statements
	"PLR2004", # Magic value used instead of constant
	"PLW0603", # Using the global statement
	"PLW2901", # redefined-loop-name
	"PLR1714", # consider-using-in
	"PLR5501", # else-if-used
]
fixable = ["ALL"]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools.packages.find]
include = ["electronicparsers*"]

[tool.setuptools.package-data]
electronicparsers = ['*/metadata.yaml', '*/README.md']

[tool.mypy]
strict = false
ignore_missing_imports = true
follow_imports = "silent"
no_strict_optional = true
disable_error_code = "import, annotation-unchecked"
